+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- fix: Publish `react` module as ES5, for backwards compatibility
 
 ## 1.8.0